+++ resources/lang/en/fields_types.php
@@ -3,31 +3,31 @@
 return [
     'checkbox_list' => [
         'title' => 'Checkbox List',
-        'description' => 'checkbox items from data source',
+        'description' => 'Checkbox items from data source',
     ],
     'color_picker' => [
         'title' => 'Color Picker',
-        'description' => 'pick a color with rgb, rgba or hsl',
+        'description' => 'Pick a color with RGB, RGBA or HSL',
     ],
     'date_picker' => [
         'title' => 'Date Picker',
-        'description' => 'full date picker',
+        'description' => 'Full date picker',
     ],
     'date_time_picker' => [
         'title' => 'Date Time Picker',
-        'description' => 'full date and time picker',
+        'description' => 'Full date and time picker',
     ],
     'file_upload' => [
         'title' => 'File Upload',
-        'description' => 'single or multiple file uploader',
+        'description' => 'Single or multiple file uploader',
     ],
     'paragraph' => [
         'title' => 'Paragraph',
-        'description' => 'display a paragraph in your form',
+        'description' => 'Display a paragraph in your form',
     ],
     'radio' => [
         'title' => 'Radio',
-        'description' => 'single choice from a datasource',
+        'description' => 'Single choice from a Datasource',
     ],
     'rich_editor' => [
         'title' => 'Rich Editor',
@@ -35,22 +35,22 @@ return [
     ],
     'select' => [
         'title' => 'Select Menu',
-        'description' => 'select single or multiple items from a dropdown list',
+        'description' => 'Select single or multiple items from a dropdown list',
     ],
     'textarea' => [
         'title' => 'Textarea',
-        'description' => 'multi line textarea',
+        'description' => 'Multi line textarea',
     ],
     'text_input' => [
         'title' => 'Text Input',
-        'description' => 'text input',
+        'description' => 'Text input',
     ],
     'time_picker' => [
         'title' => 'Time Picker',
-        'description' => 'time picker',
+        'description' => 'Time picker',
     ],
     'toggle' => [
         'title' => 'Toggle',
-        'description' => 'toggle',
+        'description' => 'Toggle',
     ],
 ];